% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAvgCalReflectance.R
\name{getAvgCalReflectance}
\alias{getAvgCalReflectance}
\title{Calculate the average calibration panel reflectance.}
\usage{
getAvgCalReflectance(caldata)
}
\arguments{
\item{caldata}{A dataframe of the calibration data from all panel scans.}
}
\value{
avgrefldata A dataframe with three columns: wavelength in nanometers,
average reflectance, and standard deviation.
}
\description{
Calculate the average calibration panel reflectance.
}

