% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getDataFromSPU.R
\name{getDataFromSPU}
\alias{getDataFromSPU}
\title{Get data from an spu file}
\usage{
getDataFromSPU(filepath, chan)
}
\arguments{
\item{filepath}{The filepath for the given files.}

\item{chan}{The channel (A or B) used for the upward facing fiber optic.}
}
\value{
rawdata A data frame with three columns: wavelength, radiance, and irradiance.
 If the upward channel is not A (the default), then the program will reverse the radiance and
 irradiance column names.
}
\description{
Get data from an spu file
}

