% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getIndex.R
\name{getIndex}
\alias{getIndex}
\title{Return a dataframe of calculated indices for each given location}
\usage{
getIndex(indexdf, indexno)
}
\arguments{
\item{indexdf}{A dataframe of index information.}

\item{indexno}{A number signifying which index in the data frame to evaluate.}
}
\value{
indexdata A dataframe of calculated indices by location
}
\description{
Return a dataframe of calculated indices for each given location
}

