% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getInterpNormRefl.R
\name{getInterpNormRefl}
\alias{getInterpNormRefl}
\title{Interpolate radiance and irradiance values for a single event file.}
\usage{
getInterpNormRefl(eventdata, calreflavgs, interpolation, limitNR)
}
\arguments{
\item{eventdata}{A data frame of a single event data file.}

\item{calreflavgs}{A data frame of the calibration panel reflectance averages.}

\item{interpolation}{One of three interpolation types: linear, spline, or cubic. 'linear'
will call the 'linear' method of the approx() function in the stats package. 'spline' or
'cubic' will respectively return the 'natural' or 'fmm' (Forsythe, Malcolm and Moler)
method for the spline() function from the stats package.}
}
\value{
interpdata A data frame of interpolated normalized reflectance values for
 consecutive whole number wavelength values from the default start value (303 nm)
 to end value (1147 nm). The data frame has four columns:wavelength, irradiance,
 radiance, and normrefl (normal reflectance).
}
\description{
Interpolate radiance and irradiance values for a single event file.
}
\seealso{
\code{\link[stats]{approx}}, \code{\link[stats]{spline}}
}

