% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getIntTime.R
\name{getIntTime}
\alias{getIntTime}
\title{Get integration time from UniSpec}
\usage{
getIntTime(inputfiles)
}
\arguments{
\item{inputfiles}{A vector of UniSpec dataset filenames from user input.}
}
\value{
inttime A string indicating the UniSpec software name and version.
}
\description{
Get integration time from UniSpec
}

