% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getNumScans.R
\name{getNumScans}
\alias{getNumScans}
\title{Get number of scans per file.}
\usage{
getNumScans(inputfiles)
}
\arguments{
\item{inputfiles}{A vector of UniSpec dataset filenames from user input.}
}
\value{
scans An integer value indicating the number of scans done and aggregated per file.
}
\description{
Get number of scans per file.
}

