% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSoftVer.R
\name{getSoftVer}
\alias{getSoftVer}
\title{Get software version info.}
\usage{
getSoftVer(inputfiles)
}
\arguments{
\item{inputfiles}{A vector of UniSpec dataset filenames from user input.}
}
\value{
software A string indicating the UniSpec software name and version.
}
\description{
Get software version info.
}

