% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getTimestamp.R
\name{getTimestamp}
\alias{getTimestamp}
\title{Get timestamp from a spu file.}
\usage{
getTimestamp(inputfiles)
}
\arguments{
\item{inputfiles}{A vector of UniSpec dataset filenames from user input.}
}
\value{
A timestamp in POSIXct format.
}
\description{
Get timestamp from a spu file.
}

